renamed missed files and fixed some issues
