+++ .github/workflows/backend-validate.yaml
@@ -55,11 +55,13 @@ jobs:
 
       - name: 🔎 Vet Code
         shell: bash
-        run: go vet backend/...
+        run: go vet ./...
+        working-directory: backend
 
       - name: 🚦 Format Code
         shell: bash
-        run: go fmt backend/...
+        run: go fmt ./...
+        working-directory: backend
 
       - name: 🔄 Auto-commit formatting
         uses: stefanzweifel/git-auto-commit-action@v5
