+++ .github/workflows/mcp-validate.yaml
@@ -76,7 +76,7 @@ jobs:
         uses: stefanzweifel/git-auto-commit-action@v5
         with:
           commit_message: "chore(fmt): apply ruff formatting"
-          file_pattern: bff/**
+          file_pattern: mcp-server/**
 
       - name: 🚦 Lint, Type Check, Security (mypy, bandit)
         shell: bash
