+++ deployments/dev/1-services/bff.tf
@@ -31,8 +31,10 @@ resource "google_cloud_run_v2_service" "bff" {
 
   template {
     service_account                  = var.iam.swim_gen_frontend.email
+    session_affinity                 = true
     max_instance_request_concurrency = 200
     timeout                          = 600
+
     containers {
       image = data.google_artifact_registry_docker_image.bff_image.self_link
       resources {
