+++ mcp-server/src/swim_gen_mcp/auth.py
