+++ mcp-server/src/swim_gen_mcp/schemas.py
